+++ common/src/main/java/net/frozenblock/wilderwild/client/renderer/blockentity/HangingTendrilRenderer.java
@@ -23,6 +23,7 @@ import net.frozenblock.wilderwild.block.HangingTendrilBlock;
 import net.frozenblock.wilderwild.block.entity.HangingTendrilBlockEntity;
 import net.frozenblock.wilderwild.client.WWModelLayers;
 import net.frozenblock.wilderwild.client.renderer.blockentity.state.HangingTendrilRenderState;
+import net.frozenblock.wilderwild.config.WWBlockConfig;
 import net.mehvahdjukaar.candlelight.api.ClientOnly;
 import net.minecraft.client.model.geom.ModelPart;
 import net.minecraft.client.renderer.blockentity.BlockEntityRendererProvider.Context;
@@ -83,5 +84,10 @@ public class HangingTendrilRenderer<T extends HangingTendrilBlockEntity> extends
 		} else {
 			renderState.sprite = this.defaultSprite;
 		}
+	}
+
+	@Override
+	public boolean shouldRender(T blockEntity, Vec3 cameraPosition) {
+		return WWBlockConfig.BILLBOARD_TENDRILS.get() && super.shouldRender(blockEntity, cameraPosition);
 	}
 }